fix: SwiftData Migration
SwiftData performs poorly when using a struct with a property that has the same name as the model's own propery. This issue is very specific to the "name" key. During migration, it should create a table with columns named "ZNAME" and "ZNAME1" for the track and album, respectively. However, it attempts to create two "ZNAME" columns, leading to a database error. As a workaround, we need to rename the property to "albumName". Good job, Apple.