fix: don't do schema migration
I really dislike this approach, but SwiftData Migration is too buggy to use at the moment, so I have to resort to this. This solution will work because the taskIdentifier is never negative. Therefore, we can use the "< 0" part to encode the file type after download and check the download status using if it's positive.