fix: debounce albumsView and put reset into load function to prevent data racing
This is an interesting one. The sort state can be changed when the view is not loaded, when the view is loaded with a changed sort state, the onChange handler will trigger multiple times and it causes data racing if the rest is not in the load callback.
This will also cause duplicated items. A easy way to trigger this is to open the track list view in offline tab(in the compact view), then go to online tab's tracks list, change the sort method, then switch back to offline tab.
The downloaded items will appear duplicated.