feat: make spotlight work with large libraries
This will batch the indexing with 250 items a time, and have confirmed with the backend developers that this is a safe number. Also reduced the cover size for indexing, as the spotlight does not need full size image which consumes more bandwidth and storage. This implementation can also resume from an incompleted indexing.