feat: Support setting default privacy setting for new playlists (#75)
* Add support for default privacy setting for new playlists

* don't overwrite public playlist if unspported by server

* Refactor playlist creation logic to handle unsupported servers

This commit refactors the logic for creating a new playlist in the PlaylistAddSheet view. It introduces a check for older Jellyfin servers that do not support private playlists. In such cases, the default value for the `publicPlaylist` state variable is set to `true` to avoid user confusion. For servers that do support shared playlists, the default value is determined by the user's preference. This change improves the user experience and ensures consistent behavior across different server versions.